+++ backend/src/main/java/develup/api/auth/Auth.java
@@ -1,4 +1,4 @@
-package develup.api;
+package develup.api.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
