+++ backend/src/main/java/develup/api/auth/AuthArgumentResolver.java
@@ -1,4 +1,4 @@
-package develup.api;
+package develup.api.auth;
 
 import static java.util.Objects.requireNonNull;
